Set default datastore adapter to `sails-disk` (string), and make sure it's loaded last.
This allows the default adapter to use an already-loaded sails-disk if there is one (because another datastore loaded it), falling back to using the bundled version as a last resort.  This avoids an error that can come up when using a second datastore with `sails-disk`, because it may load _both_ the bundled and local versions, which presents a conflict (can't have two different adapter modules with the same identity).

fixes https://trello.com/c/r7TfWyEU/211-bug-sh-orm200-16-should-be-able-to-have-two-or-more-datastores-that-share-the-same-adapter